+++ packages/core/components/auditing.js
@@ -6,6 +6,7 @@
  * You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
+const { EOL } = require( "node:os" )
 const _ = require( "lodash" );
 const tools = require( "#tools" );
 const logger = require( "#logger" );
@@ -132,7 +133,7 @@ class Auditing {
         } else {
             console.log( Auditing.#formatConsoleMessage( logEntry ) );
             if ( !_.isEmpty( logEntry.data ) ) {
-                console.log( `   » ${ Auditing.#formatConsoleData( logEntry ) }` );
+                console.log( Auditing.#formatConsoleData( logEntry.data, "   " ) );
             }
         }
     }
@@ -154,12 +155,53 @@ class Auditing {
      * Used to format a log entry data payload for the Node console.
      *
      * @method
-     * @param {TiLogEntry} logEntry
+     * @param {*} data
+     * @param {string} [prefix=""]
+     * @param {number} [currentDepth=0]
+     * @param {number} [maxDepth=5]
+     * @param {Set} [visited=new Set()]
      * @returns {string}
      * @private
      */
-    static #formatConsoleData( logEntry ) {
-        return tools.stringifyJSON( logEntry.data );
+    static #formatConsoleData( data, prefix = "", currentDepth = 0, maxDepth = 5, visited = new Set() ) {
+        if ( data === null || data === undefined || !_.isObjectLike( data ) ) {
+            // Handle null, undefined, and primitive values:
+            return prefix + "» " + String( data );
+        } else if ( currentDepth >= maxDepth ) {
+            // Check max depth:
+            return prefix + "! [max data depth reached]";
+        } else if ( visited.has( data ) ) {
+            // Check for circular references:
+            return prefix + "! [circular reference detected]";
+        } else {
+            let formattedData = "";
+            visited.add( data );
+
+            _.forOwn( data, ( value, key ) => {
+                let stackLines = ( _.isString( value ) ) ? value.split( "\n" ) : [ value ];
+                if ( stackLines.length > 1 ) {
+                    _.forEach( stackLines, ( line, idx ) => {
+                        if ( idx === 0 ) {
+                            formattedData += prefix + `» ${ key }: ${ _.trim( line ) }` + EOL;
+                        } else {
+                            formattedData += prefix + `- ${ _.trim( line ) }` + EOL;
+                        }
+                    } );
+                } else {
+                    if ( _.isObjectLike( value ) ) {
+                        formattedData += prefix + `» ${ key }:` + EOL + Auditing.#formatConsoleData( value, "   " + prefix, currentDepth + 1, maxDepth, visited ) + EOL;
+                    } else {
+                        formattedData += prefix + `» ${ key }: ${ value }` + EOL;
+                    }
+                }
+            } );
+
+            if ( formattedData.endsWith( EOL ) ) {
+                formattedData = formattedData.slice( 0, -EOL.length );
+            }
+
+            return formattedData;
+        }
     }
 }